Create Indian Education System_EDA
Exploratory Data Analysis is performed on the dataset containing information such as dropout rates, percentage-of-schools-with-water-facility, schools-with-girls-toilet, gross-enrollment-ratio, percentage-of-schools-with-comps, schools-with-boys-toilet, percentage-of-schools-with-electricity. The data varies between the year 2012 and 2016. The dataset is taken from kaggle(https://www.kaggle.com/datasets/vidyapb/indian-school-education-statistics?resource=download). The data is taken for analysis purpose in order to study and derive conclusions.